// src/grids/src/grid/base/interface.ts
export interface IFilterMUI {
    create?: void | Function | string;
    read?: Object | Function | string;
    write?: void | Function | string;
}
// ...
export interface EJ2Intance extends HTMLElement {
    // eslint-disable-next-line
    ej2_instances: Object | Object[];
}
// ...
export interface IFilterCreate {
    column?: Column;
    target?: HTMLElement;
    getOptrInstance?: FlMenuOptrUI;
    localizeText?: L10n;
    dialogObj?: Dialog;
}